Revert "Disable context processors when rendering e-mail"
Django internally caches the list of functions specified in
TEMPLATE_CONTEXT_PROCESSORS, to save on imports:

https://github.com/django/django/blame/1.4/django/template/context.py#L139

Side effect of the reverted commit: If a comm panel preview is the first
request handled after a code reload, the website will run with no
context processors.

With a testing procedure of sitting on the preview page and refreshing,
this gives the appearance of that fix having worked.

(As MessageRequest.parseSmartText was only ever triggered by the dbmail
cron script, half of this didn't actually need to be reverted for
correctness.)

This reverts commit d9263ef802fa4abc351a69e0425562fd059623e6.